Module Acutis.Ast

Define the untyped abstract syntax tree to be generated by the Parser.

module Dict : sig ... end

This is a Map.String which enforces that keys cannot change after they're set.

module Record : sig ... end

module Interface : sig ... end

module Pattern : sig ... end

type trim =

| No_trim

| Trim

type escape =

| No_escape

| Escape

type echo =

| Ech_var of Loc.t * string * escape

| Ech_component of Loc.t * string

| Ech_string of Loc.t * string

type node =

| Text of string * trim * trim

| Echo of echo list * echo

| Match of Loc.t * Pattern.t Nonempty.t * case Nonempty.t

| Map_list of Loc.t * Pattern.t * case Nonempty.t

| Map_dict of Loc.t * Pattern.t * case Nonempty.t

| Component of Loc.t * string * string * Pattern.t Dict.t * child Dict.t

| Interface of Loc.t * Interface.t list

and case = {

pats : (Loc.t * Pattern.t Nonempty.t) Nonempty.t;

nodes : t;

}

and child =

| Child_name of Loc.t * string

| Child_block of t

and t = node list

val equal : t -> t -> bool

val pp : Stdlib.Format.formatter -> t -> unit