Module Acutis.Compile

Orchestrate the Lexer, Parser, Typechecker, and Matching to produce the final template.

val parse : fname:string -> Stdlib.Lexing.lexbuf -> Ast.t

raises Error.Acutis_error

type escape = Ast.escape =

| No_escape

| Escape

type echo = Typechecker.echo =

| Ech_var of string * escape

| Ech_string of string

type 'a node =

| Text of string

| Echo of (string * escape) list * echo

| Match of 'a data Data.t array * 'a nodes Matching.t

| Map_list of 'a data Data.t * 'a nodes Matching.t

| Map_dict of 'a data Data.t * 'a nodes Matching.t

| Component of 'a * 'a data Data.t Map.String.t

The names of variables are preserved as strings in the Data.t values.

and 'a data =

| Var of string

| Block of 'a nodes

and 'a nodes = 'a node list

type 'a template =

| Src of 'a template nodes

| Fun of Typescheme.t Map.String.t * 'a

type 'a t = {

types : Typescheme.t Map.String.t;

nodes : 'a template nodes;

name : string;

}

module Components : sig ... end

val make : fname:string -> 'a Components.t -> Stdlib.Lexing.lexbuf -> 'a t

val from_string : fname:string -> 'a Components.t -> string -> 'a t

val from_channel : fname:string -> 'a Components.t -> Stdlib.in_channel -> 'a t