Module Acutis.Dagmap

A utility to help Typechecker and Compile enforce that templates form a directed acyclic graph.

type ('a, 'b) t

'a is the type of values before linking into the graph. 'b is the type of values after they've been linked.

val make : f:( ( 'a, 'b ) t -> 'a -> 'b ) -> ?root:string -> 'a Map.String.t -> ( 'a, 'b ) t

Use get inside the f function.

parameter f
Transforms 'a to 'b.

parameter root
The root component, if one is defined.

val get : string -> ( 'a, 'b ) t -> 'b

Gets the 'b value associated with a string key.

raises Error.Acutis_error
When a key does not exist or if the graph forms a cycle.

val prelinked : string -> 'a Map.String.t -> ( 'b, 'a ) t

val link_all : ( 'a, 'b ) t -> 'b Map.String.t