Module Acutis.Data

The internal representation of runtime data.

module Const : sig ... end

type 'a t =

| Other of 'a	(*
At runtime, this stores any raw input with an unknown type. In a compiled template, this stores the names of variables.
*)

| Nil	(*
Represents both null and [].
*)

| Array of 'a t array	(*
Tuples are compiled to arrays.
*)

| Dict of 'a t Map.String.t	(*
Records and dictionaries are compiled to string maps.
*)

| Const of Const.t * Typescheme.Variant.extra	(*
Integers, strings, and floats must be wrapped to accommodate the polymorphic echo.
*)

The boxed representation of runtime data.

Constructing data

val other : 'a -> 'a t

val null : _ t

val const : Const.t -> Typescheme.Variant.extra -> _ t

val int : int -> _ t

val bool : int -> _ t

val string : string -> _ t

val float : float -> _ t

val some : 'a t -> 'a t

val dict : 'a t Map.String.t -> 'a t

val tuple : 'a t array -> 'a t

val list_cons : 'a t -> 'a t -> 'a t

val list_rev : 'a t -> 'a t

val list_empty : _ t

val flat_map : ( 'a -> 'b t ) -> 'a t -> 'b t

Transform the contents of Unknown values to a different type of data.

Deconstructing data

val get_const : _ t -> Const.t

val get_tuple : 'a t -> 'a t array

val get_dict : 'a t -> 'a t Map.String.t

val is_null : _ t -> bool

val get_nullable : 'a t -> 'a t option

val fold_list : ( index:'a t -> 'b -> 'a t -> 'b ) -> 'b -> 'a t -> 'b

val fold_dict : ( index:'a t -> 'b -> 'a t -> 'b ) -> 'b -> 'a t -> 'b

Echoing data

val echo : _ t -> string

This only works on constants.