Parameter Make.I

Define the semantics of our abstract language.

Statements and expressions.

type 'a stmt

A statement, generally some side-effecting code.

type 'a obs

The final evaluation result.

val observe : 'a stmt -> 'a obs

Observe the evaluation result after all transformations have applied.

val (|:) : unit stmt -> 'a stmt -> 'a stmt

Sequence statements. Evaluation order is unspecified in OCaml. To guarantee that statements are evaluated in your intended order, you must let-bind each one before applying them to |:.

type 'a exp

An expression.

val return : 'a exp -> 'a stmt

Return value 'a from a function.

val let$ : (string * 'a exp) -> ('a exp -> 'b stmt) -> 'b stmt

Define a new immutable binding. The string is used for pretty-printing.

type 'a mut

A mutable reference variable.

val let& : (string * 'a exp) -> ('a mut -> 'b stmt) -> 'b stmt

Define a new reference variable. The string is used for pretty-printing.

val deref : 'a mut -> 'a exp

val (:=) : 'a mut -> 'a exp -> unit stmt

val incr : int mut -> unit stmt

Functions.

val lambda : ('a exp -> 'b stmt) -> ('a -> 'b) exp

val (@@) : ('a -> 'b) exp -> 'a exp -> 'b exp

Control flow.

val if_ : bool exp -> then_:(unit -> unit stmt) -> else_:(unit -> unit stmt) option -> unit stmt

val while_ : (unit -> bool exp) -> (unit -> unit stmt) -> unit stmt

type external_data

Data from the outside world that we need to decode.

type 'a promise

An asynchronous monad.

type import

Information to import a function from external code.

val import : import -> ((external_data -> string promise) exp -> 'a stmt) -> 'a stmt

The import semantics are abstract because runtimes vary.

val export : 'a exp -> 'a stmt

Standard values.

val unit : unit stmt

val not : bool exp -> bool exp

val int : int -> int exp

val float : float -> float exp

val string : string -> string exp

val bool : bool -> bool exp

val pair : ('a exp * 'b exp) -> ('a * 'b) exp

val equal_int : int exp -> int exp -> bool exp

val equal_string : string exp -> string exp -> bool exp

val int_to_string : int exp -> string exp

val int_to_float : int exp -> float exp

val float_to_string : float exp -> string exp

val bool_to_string : int exp -> string exp

Arrays.

val array : 'a exp array -> 'a array exp

val array_init : int exp -> 'a exp -> 'a array exp

val (.%()) : 'a array exp -> int exp -> 'a exp

val (.%()<-) : 'a array exp -> int exp -> 'a exp -> unit stmt

val array_concat : string array exp -> string exp -> string exp

Hash tables.

type 'a hashtbl

A mutable map of string keys to 'a values.

val hashtbl : (string * 'a) exp Stdlib.Seq.t -> 'a hashtbl exp

val hashtbl_create : unit -> 'a hashtbl exp

val (.%{}) : 'a hashtbl exp -> string exp -> 'a exp

val (.%{}<-) : 'a hashtbl exp -> string exp -> 'a exp -> unit stmt

val hashtbl_mem : 'a hashtbl exp -> string exp -> bool exp

val hashtbl_copy : 'a hashtbl exp -> 'a hashtbl exp

val hashtbl_iter : 'a hashtbl exp -> (string exp -> 'a exp -> unit stmt) -> unit stmt

Promises.

val promise : 'a exp -> 'a promise exp

val bind_array : string promise array exp -> (string array -> 'a promise) exp -> 'a promise exp

Buffers.

type buffer

This type is not a typical "buffer" since it must work with promises. The language may implement it however is most suitable.

val buffer_create : unit -> buffer exp

val buffer_add_string : buffer exp -> string exp -> unit stmt

val buffer_add_promise : buffer exp -> string promise exp -> unit stmt

These are lambdas to minimize generated code.

val buffer_to_promise : (buffer -> string promise) exp

val escape : (string -> string) exp

Mutable stacks.

type 'a stack

val stack_create : unit -> 'a stack exp

val stack_is_empty : 'a stack exp -> bool exp

val stack_push : 'a stack exp -> 'a exp -> unit stmt

val stack_drop : 'a stack exp -> unit stmt

val stack_concat : string stack exp -> string exp -> string exp

Raise errors.

type error

val raise : error exp -> 'a stmt

val error : string exp -> error exp

Runtime data.

type data

Boxed runtime data. Either a string, an integer, a float, an array, or a hash table.

module Data : sig ... end

module External : sig ... end

The foreign data before it's parsed into Data.t.