Module Instruct.MakeTrans

Apply a transformation module and a semantics module to produce a new semantics module that uses the transformation state.

The module this creates won't do any transformations by itself, and its values are defined as identity functions. You need to override specific functions to apply transformations. This is used in PrintJs to optimize its output.

Define the semantics of our abstract language.

Parameters

module T : TRANS

module F : SEM with type 'a exp = 'a T.from_exp and type 'a stmt = 'a T.from_stmt

Signature

Statements and expressions.

type 'a stmt = 'a T.stmt

A statement, generally some side-effecting code.

type 'a obs = 'a F.obs

The final evaluation result.

val observe : 'a stmt -> 'a obs

Observe the evaluation result after all transformations have applied.

val (|:) : unit stmt -> 'a stmt -> 'a stmt

Sequence statements. Evaluation order is unspecified in OCaml. To guarantee that statements are evaluated in your intended order, you must let-bind each one before applying them to |:.

type 'a exp = 'a T.exp

An expression.

val return : 'a exp -> 'a stmt

Return value 'a from a function.

val stmt : 'a exp -> 'a stmt

val let$ : (string * 'a exp) -> ('a exp -> 'b stmt) -> 'b stmt

Define a new immutable binding. The string is used for pretty-printing.

type 'a mut = 'a F.mut

A mutable reference variable.

val let& : (string * 'a exp) -> ('a mut -> 'b stmt) -> 'b stmt

Define a new reference variable. The string is used for pretty-printing.

val deref : 'a mut -> 'a exp

val (:=) : 'a mut -> 'a exp -> unit stmt

val incr : int mut -> unit stmt

Functions.

val lambda : ('a exp -> 'b stmt) -> ('a -> 'b) exp

val (@@) : ('a -> 'b) exp -> 'a exp -> 'b exp

Control flow.

val if_ : bool exp -> then_:(unit -> unit stmt) -> unit stmt

val if_else : bool exp -> then_:(unit -> 'a stmt) -> else_:(unit -> 'a stmt) -> 'a stmt

val while_ : ('a exp -> bool exp) -> 'a mut -> (unit -> unit stmt) -> unit stmt

Standard values.

val unit : unit stmt

val not : bool exp -> bool exp

val int : int -> int exp

val float : float -> float exp

val string : string -> string exp

val bool : bool -> bool exp

val equal_int : int exp -> int exp -> bool exp

val equal_string : string exp -> string exp -> bool exp

val string_of_int : int exp -> string exp

val float_of_int : int exp -> float exp

val string_of_float : float exp -> string exp

val string_of_bool : int exp -> string exp

Arrays.

val array : 'a exp array -> 'a array exp

val array_make : int exp -> 'a exp -> 'a array exp

val (.%()) : 'a array exp -> int exp -> 'a exp

val (.%()<-) : 'a array exp -> int exp -> 'a exp -> unit stmt

Hash tables.

type 'a hashtbl = 'a F.hashtbl

A mutable map of string keys to 'a values.

val hashtbl : (string exp * 'a exp) Stdlib.Seq.t -> 'a hashtbl exp

val hashtbl_create : unit -> 'a hashtbl exp

val (.%{}) : 'a hashtbl exp -> string exp -> 'a exp

val (.%{}<-) : 'a hashtbl exp -> string exp -> 'a exp -> unit stmt

val hashtbl_mem : 'a hashtbl exp -> string exp -> bool exp

val hashtbl_copy : 'a hashtbl exp -> 'a hashtbl exp

val hashtbl_iter : 'a hashtbl exp -> (string exp -> 'a exp -> unit stmt) -> unit stmt

Buffers.

type buffer = F.buffer

A mutable string buffer.

val buffer_create : unit -> buffer exp

val buffer_add_string : buffer exp -> string exp -> unit stmt

val buffer_add_buffer : buffer exp -> buffer exp -> unit stmt

val buffer_add_escape : buffer exp -> string exp -> unit stmt

E.g. sanitize HTML syntax.

val buffer_contents : buffer exp -> string exp

val buffer_length : buffer exp -> int exp

val buffer_clear : buffer exp -> unit stmt

Promises.

type 'a promise = 'a F.promise

An asynchronous monad.

val promise : 'a exp -> 'a promise exp

val bind : 'a promise exp -> ('a -> 'b promise) exp -> 'b promise exp

val error : string exp -> 'a promise exp

Importing and exporting.

type external_data = F.external_data

Data from the outside world that we need to decode.

type import = F.import

Information to import a function from external code.

val import : import -> ((external_data -> string promise) exp -> 'a stmt) -> 'a stmt

val export : 'a exp -> 'a stmt

Runtime data.

type data = F.data

Runtime data. Either a string, an integer, a float, an array, or a hash table.

module Data : sig ... end

module External : sig ... end

The foreign data before it's parsed into Data.t.