Module Instruct.SymT

Apply a transformation module and a symantics module to produce a new symantics module that uses the transformation state.

The module this creates won't do any transformations by itself, and its values are defined as identity functions. You need to override specific functions with optimized forms.

Parameters

module T : TRANS

module F : SYM with type 'a exp = 'a T.from_exp and type 'a stmt = 'a T.from_stmt

Signature

type 'a exp = 'a T.exp

type 'a stmt = 'a T.stmt

type 'a obs = 'a F.obs

val observe : 'a T.stmt -> 'b F.obs

val (|:) : unit T.stmt -> 'a T.stmt -> 'b T.stmt

val return : 'a T.exp -> 'b T.stmt

val let$ : (string * 'a T.exp) -> ('b T.exp -> 'c T.stmt) -> 'd T.stmt

type 'a mut = 'a F.mut

val let& : (string * 'a T.exp) -> ('b F.mut -> 'c T.stmt) -> 'd T.stmt

val deref : 'a F.mut -> 'b T.exp

val (:=) : 'a F.mut -> 'b T.exp -> unit T.stmt

val incr : int F.mut -> unit T.stmt

val lambda : ('a T.exp -> 'b T.stmt) -> ('c -> 'd) T.exp

val (@@) : ('a -> 'b) T.exp -> 'c T.exp -> 'd T.exp

val if_ : bool T.exp -> then_:(unit -> unit T.stmt) -> else_:(unit -> unit T.stmt) option -> unit T.stmt

val while_ : (unit -> bool T.exp) -> (unit -> unit T.stmt) -> unit T.stmt

type external_data = F.external_data

type 'a promise = 'a F.promise

type import = F.import

val import : F.import -> ((F.external_data -> string F.promise) T.exp -> 'a T.stmt) -> 'b T.stmt

val export : 'a T.exp -> 'b T.stmt

val unit : unit T.stmt

val not : bool T.exp -> bool T.exp

val int : int -> int T.exp

val float : float -> float T.exp

val string : string -> string T.exp

val bool : bool -> bool T.exp

val pair : ('a T.exp * 'b T.exp) -> ('c * 'd) T.exp

val equal_int : int T.exp -> int T.exp -> bool T.exp

val equal_string : string T.exp -> string T.exp -> bool T.exp

val int_to_string : int T.exp -> string T.exp

val int_to_float : int T.exp -> float T.exp

val float_to_string : float T.exp -> string T.exp

val bool_to_string : int T.exp -> string T.exp

val array : 'a T.exp array -> 'b array T.exp

val array_init : int T.exp -> 'a T.exp -> 'b array T.exp

val (.%()) : 'a array T.exp -> int T.exp -> 'b T.exp

val (.%()<-) : 'a array T.exp -> int T.exp -> 'b T.exp -> unit T.stmt

val array_concat : string array T.exp -> string T.exp -> string T.exp

type 'a hashtbl = 'a F.hashtbl

val hashtbl : (string * 'a) T.exp Stdlib.Seq.t -> 'a F.hashtbl T.exp

val hashtbl_create : unit -> 'a F.hashtbl T.exp

val (.%{}) : 'a F.hashtbl T.exp -> string T.exp -> 'b T.exp

val (.%{}<-) : 'a F.hashtbl T.exp -> string T.exp -> 'b T.exp -> unit T.stmt

val hashtbl_mem : 'a F.hashtbl T.exp -> string T.exp -> bool T.exp

val hashtbl_copy : 'a F.hashtbl T.exp -> 'a F.hashtbl T.exp

val hashtbl_iter : 'a F.hashtbl T.exp -> (string T.exp -> 'b T.exp -> unit T.stmt) -> unit T.stmt

val promise : 'a T.exp -> 'b F.promise T.exp

val bind_array : string F.promise array T.exp -> (string array -> 'a F.promise) T.exp -> 'a F.promise T.exp

type buffer = F.buffer

val buffer_create : unit -> F.buffer T.exp

val buffer_add_string : F.buffer T.exp -> string T.exp -> unit T.stmt

val buffer_add_promise : F.buffer T.exp -> string F.promise T.exp -> unit T.stmt

val buffer_to_promise : (F.buffer -> string F.promise) T.exp

val escape : (string -> string) T.exp

type 'a stack = 'a F.stack

val stack_create : unit -> 'a F.stack T.exp

val stack_is_empty : 'a F.stack T.exp -> bool T.exp

val stack_push : 'a F.stack T.exp -> 'b T.exp -> unit T.stmt

val stack_drop : 'a F.stack T.exp -> unit T.stmt

val stack_concat : string F.stack T.exp -> string T.exp -> string T.exp

type error = F.error

val raise : F.error T.exp -> 'a T.stmt

val error : string T.exp -> F.error T.exp

type data = F.data

module Data : sig ... end

module External : sig ... end