Module Acutis.Lexer

The lexer module generated by ocamllex.

exception Error

type state

The lexer relies on different rules depending on the type of expression it's currently in (echo, text, etc). This state type tracks that as an addition to the regular lexbuf state.

val make_state : unit -> state

val acutis : state -> Stdlib.Lexing.lexbuf -> Parser.token