Module Matching.Exit

Each "exit" is given an integer key which we can use to look up the AST nodes to follow after the tree. We use integers because exits can be copied when trees merge, and we don't want to duplicate entire ASTs.

type key

Internally: int

type 'a t

val get : 'a t -> key -> 'a

val map : ('a -> 'b) -> 'a t -> 'b t

val key_to_int : key -> int

val to_seqi : 'a t -> (key * 'a) Stdlib.Seq.t