Module Matching.Exits

Each "exit" is given a key which we can use to look up the AST nodes to follow after the tree. We use keys because exits can be copied when trees merge, and we don't want to duplicate entire trees.

type key

val key_to_int : key -> int

type 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

val binding_exists : _ t -> bool

val nodes : 'a t -> 'a Stdlib.Seq.t

val to_seq : 'a t -> (key * string list * 'a) Stdlib.Seq.t

This returns each exit's key, list of binding names, and nodes.