Module Acutis.Render

The main runtime for executing templates.

module type MONAD = sig ... end

Your standard monad type.

module type DATA = sig ... end

Decode and encode input data.

module type S = sig ... end

Output signature for Make.

module Make (M : MONAD) (D : DATA) : S with type t = string M.t and type data = D.t

A functor that builds an implementation for a given monadic output type and a given data input type.

module MakeString (D : DATA) : S with type t = string and type data = D.t

A simpler version of Make that only requires a data module and outputs a string.