Module Render.Make

A functor that builds an implementation for a given concurrency interface and a given decodable input type.

Output signature for Make.

Parameters

module C : CONCURRENT

module D : DECODABLE

Signature

type t = string C.promise

type data = D.t

val eval : (data -> t) Compile.t -> data -> t

Apply data to a template and return the rendered output.