Parameter Make.D

Decode and encode input data.

module Linear : sig ... end

A linear container such as a list or array.

module Assoc : sig ... end

A key-value container such as an association list, a string map, etc.

type t

Decoding

val classify : t -> [ `Null | `Bool of bool | `Int of int | `Float of float | `String of string | `Assoc of t Assoc.t | `List of t Linear.t ]

Encoding

val null : t

val some : t -> t

val of_float : float -> t

val of_string : string -> t

val of_bool : bool -> t

val of_int : int -> t

val of_seq : t Stdlib.Seq.t -> t

val of_map : t Map.String.t -> t

Debugging

val pp : Stdlib.Format.formatter -> t -> unit