Module type Render.DECODABLE

Decode and encode input data.

Container types

type 'a linear

A linear container such as a list or array.

val length : 'a linear -> int

val iteri : (int -> 'a -> unit) -> 'a linear -> unit

type 'a assoc

A key-value container such as an association list or a string map.

val assoc_find : string -> 'a assoc -> 'a

val assoc_mem : string -> 'a assoc -> bool

val assoc_iter : (string -> 'a -> unit) -> 'a assoc -> unit

type t

Encoding

val null : t

val some : t -> t

val of_float : float -> t

val of_string : string -> t

val of_bool : bool -> t

val of_int : int -> t

val of_array : t array -> t

val of_assoc : (string * t) Stdlib.Seq.t -> t

Decoding

val decode_int : t -> int option

val decode_string : t -> string option

val decode_float : t -> float option

val decode_bool : t -> bool option

val decode_some : t -> t option

val decode_linear : t -> t linear option

val decode_assoc : t -> t assoc option

Debugging

val to_string : t -> string