Module Acutis.Typescheme

Defines the types (integer, record, etc.).

module Variant : sig ... end

This is a common structure shared by Enum and Union.

module Enum : sig ... end

module Union : sig ... end

type ty =

| Unknown of Variant.row Stdlib.ref	(*
The row is for unification with variant types during destructuring.
*)

| Int

| Float

| String

| Nullable of t

| List of t

| Tuple of t list

| Record of t Map.String.t Stdlib.ref

| Dict of t * Set.String.t Stdlib.ref	(*
The set tracks which keys have been used, so they can build pattern-matching decision trees.
*)

| Enum of Enum.t

| Union of string * t Union.t

and t = ty Stdlib.ref

Public API for declaring type schemes.

val unknown : unit -> t

val int : unit -> t

val float : unit -> t

val string : unit -> t

val nullable : t -> t

val list : t -> t

val tuple : t list -> t

val record : (string * t) list -> t

val dict : t -> t

val enum_int : Variant.row -> int list -> t

val enum_string : Variant.row -> string list -> t

val boolean : unit -> t

val false_only : unit -> t

val true_only : unit -> t

val union_int : Variant.row -> string -> (int * (string * t) list) list -> t

val union_string : Variant.row -> string -> (string * (string * t) list) list -> t

val union_boolean : string -> f:(string * t) list -> t:(string * t) list -> t

val union_false_only : string -> (string * t) list -> t

val union_true_only : string -> (string * t) list -> t

val make : (string * t) list -> t Map.String.t

val empty : t Map.String.t

Internal utilities.

val internal_record : t Map.String.t Stdlib.ref -> t

val internal_dict_keys : t -> Set.String.t Stdlib.ref -> t

val internal_copy_record : t Map.String.t -> t Map.String.t

val internal_bool : int list -> t

val pp : Stdlib.Format.formatter -> t -> unit

val pp_interface : Stdlib.Format.formatter -> t Map.String.t -> unit

val equal : t -> t -> bool