Module Acutis_internals

module Ast : sig ... end

Define the untyped abstract syntax tree to be generated by the Parser.

module Compile : sig ... end

Orchestrate the Lexer, Parser, Typechecker, and Matching to produce the final template.

module Error : sig ... end

All of the error messages.

module Instruct : sig ... end

Define the instructions to evaluate a Compile result at runtime.

module Lexer : sig ... end

The lexer module generated by ocamllex.

module Loc : sig ... end

Define the locations of expressions in a source file.

module Matching : sig ... end

Compile patterns into decision trees.

module Nonempty : sig ... end

Lists with at least one item.

module Parser : sig ... end

module ParserMessages : sig ... end

Turn Menhir errors into human-readable messages.

module Pp : sig ... end

Helpers for Format functions.

module Sexp : sig ... end

Serialize internal data into S-expressions for debugging.

module Typechecker : sig ... end

Type-checks the untyped Ast.t and constructs a typed tree.