Module F.External

Data from the outside world that we need to decode.

type t

val null : t exp

val some : t exp -> t exp

val of_int : int exp -> t exp

val of_float : float exp -> t exp

val of_string : string exp -> t exp

val of_bool : bool exp -> t exp

val of_seq : t Stdlib.Seq.t exp -> t exp

val of_seq_assoc : (string * t) Stdlib.Seq.t exp -> t exp

type 'a assoc

A key-value container such as an association list or a string map.

val assoc_find : string exp -> 'a assoc exp -> 'a exp

val assoc_mem : string exp -> 'a assoc exp -> bool exp

val assoc_to_seq : 'a assoc exp -> (string * 'a) Stdlib.Seq.t exp

type 'a decoder

val get_int : int decoder

val get_string : string decoder

val get_float : float decoder

val get_bool : bool decoder

val get_some : t decoder

val get_seq : t Stdlib.Seq.t decoder

val get_assoc : t assoc decoder

val decode : 'a decoder -> t exp -> ok:('a exp -> 'b stm) -> error:(unit -> 'b stm) -> 'b stm

val to_string : t exp -> string exp

val marshal : 'a exp -> t exp