Module type Instruct.SEM

Define the semantics of our abstract language.

Statements and expressions.

type 'a stmt

A statement, generally some side-effecting code.

type 'a obs

The final evaluation result.

val observe : 'a stmt -> 'a obs

Observe the evaluation result after all transformations have applied.

type 'a exp

An expression.

val return : 'a exp -> 'a stmt

Return value 'a from a function.

val stmt : 'a exp -> 'a stmt

type 'a ref

A mutable reference variable.

val let| : unit stmt -> (unit -> 'a stmt) -> 'a stmt

Sequence statements.

val let$ : (string * 'a exp) -> ('a exp -> 'b stmt) -> 'b stmt

Define a new immutable binding. The string is used for pretty-printing.

val let& : (string * 'a exp) -> ('a ref -> 'b stmt) -> 'b stmt

Define a new reference variable. The string is used for pretty-printing.

val (!) : 'a ref -> 'a exp

val (:=) : 'a ref -> 'a exp -> unit stmt

val incr : int ref -> unit stmt

Functions.

val lambda : ('a exp -> 'b stmt) -> ('a -> 'b) exp

val (@@) : ('a -> 'b) exp -> 'a exp -> 'b exp

Control flow.

val if_ : bool exp -> then_:(unit -> unit stmt) -> unit stmt

val if_else : bool exp -> then_:(unit -> 'a stmt) -> else_:(unit -> 'a stmt) -> 'a stmt

val while_ : ('a exp -> bool exp) -> 'a ref -> (unit -> unit stmt) -> unit stmt

Standard values.

val unit : unit stmt

val not : bool exp -> bool exp

val int : int -> int exp

val float : float -> float exp

val string : string -> string exp

val bool : bool -> bool exp

val (=) : 'a exp -> 'a exp -> bool exp

val string_of_int : int exp -> string exp

val float_of_int : int exp -> float exp

val string_of_float : float exp -> string exp

val string_of_bool : bool exp -> string exp

Strings and characters.

val string_iter : string exp -> (char exp -> unit stmt) -> unit stmt

val match_char : char exp -> (char -> 'a stmt) -> 'a stmt

The given function applies to the following characters and it ignores all others: '&' '"' '\'' '>' '<' '/' '`' '=' .

Arrays.

val array : 'a exp array -> 'a array exp

val array_make : int exp -> 'a exp -> 'a array exp

val (.%()) : 'a array exp -> int exp -> 'a exp

val (.%()<-) : 'a array exp -> int exp -> 'a exp -> unit stmt

Hash tables.

type 'a hashtbl

A mutable map of string keys to 'a values.

val hashtbl : (string exp * 'a exp) Stdlib.Seq.t -> 'a hashtbl exp

val hashtbl_create : unit -> 'a hashtbl exp

val (.%{}) : 'a hashtbl exp -> string exp -> 'a exp

val (.%{}<-) : 'a hashtbl exp -> string exp -> 'a exp -> unit stmt

val hashtbl_mem : 'a hashtbl exp -> string exp -> bool exp

val hashtbl_iter : 'a hashtbl exp -> (string exp -> 'a exp -> unit stmt) -> unit stmt

Buffers.

type buffer

A mutable string buffer.

val buffer_create : unit -> buffer exp

val buffer_add_string : buffer exp -> string exp -> unit stmt

val buffer_add_buffer : buffer exp -> buffer exp -> unit stmt

val buffer_add_char : buffer exp -> char exp -> unit stmt

val buffer_contents : buffer exp -> string exp

val buffer_length : buffer exp -> int exp

val buffer_clear : buffer exp -> unit stmt

Promises.

type 'a promise

An asynchronous monad.

val promise : 'a exp -> 'a promise exp

val bind : 'a promise exp -> ('a -> 'b promise) exp -> 'b promise exp

val error : string exp -> 'a promise exp

Data

module External : sig ... end

Foreign data before it's parsed into Data.t.

module Data : sig ... end

Runtime data.

Importing and exporting.

type import

Information to import a function from external code.

val import : import -> ((External.t -> string promise) exp -> 'a stmt) -> 'a stmt

val export : 'a exp -> 'a stmt