Module Acutis

This is the main API wrapper around the compiler internals.

exception Acutis_error of string

This is raised with a message if any part of the process fails.

Declaring type schemes.

This is used primarily for component functions.

type ty = private Acutis_internals.Typechecker.Type.t

type typescheme = ty Stdlib.Map.Make(Stdlib.String).t

val unknown : unit -> ty

val int : unit -> ty

val float : unit -> ty

val string : unit -> ty

val nullable : ty -> ty

val list : ty -> ty

val tuple : ty list -> ty

val record : (string * ty) Stdlib.Seq.t -> ty

val dict : ty -> ty

val enum_int : [ `Open | `Closed ] -> int Stdlib.Seq.t -> ty

val enum_string : [ `Open | `Closed ] -> string Stdlib.Seq.t -> ty

val boolean : unit -> ty

val false_only : unit -> ty

val true_only : unit -> ty

val union_int : [ `Open | `Closed ] -> string -> (int * (string * ty) Stdlib.Seq.t) Stdlib.Seq.t -> ty

val union_string : [ `Open | `Closed ] -> string -> (string * (string * ty) Stdlib.Seq.t) Stdlib.Seq.t -> ty

val union_boolean : string -> f:(string * ty) Stdlib.Seq.t -> t:(string * ty) Stdlib.Seq.t -> ty

val union_false_only : string -> (string * ty) Stdlib.Seq.t -> ty

val union_true_only : string -> (string * ty) Stdlib.Seq.t -> ty

val typescheme : (string * ty) Stdlib.Seq.t -> typescheme

val typescheme_empty : typescheme

Compiling templates.

type 'a comp = private 'a Acutis_internals.Compile.Components.source

A template component. It is either from Acutis source code or it is 'a. 'a may be a function or data to load a function, depending on how the template is to be rendered.

val comp_parse : fname:string -> name:string -> Stdlib.Lexing.lexbuf -> 'a comp

Parses Acutis source into a template component but doesn't type-check yet. fname is the file path and used here for debugging. name is the name the component is called in Acutis code.

val comp_fun : name:string -> typescheme -> 'a -> 'a comp

Convert a function (or possibly data for loading a function) into a template component.

type 'a comps_compiled = private 'a Acutis_internals.Compile.Components.t

A group of compiled and linked components.

val comps_compile : 'a comp Stdlib.Seq.t -> 'a comps_compiled

Type-checks, optimizes, and links the components.

val comps_empty : 'a comps_compiled

type parsed

A template that's been parsed but not type-checked, optimized, or linked yet.

val parse : fname:string -> Stdlib.Lexing.lexbuf -> parsed

type 'a compiled = private 'a Acutis_internals.Compile.t

A completely compiled template.

val compile : 'a comps_compiled -> parsed -> 'a compiled

Type-checks, optimizes, and links the template with its components.

val compile_interface : fname:string -> Stdlib.Lexing.lexbuf -> typescheme

Parses and type-checks an interface with no template.

val get_typescheme : 'a compiled -> typescheme

Rendering templates with OCaml.

module type MONAD = sig ... end

A monad interface for async operations.

module type DECODABLE = sig ... end

Decode and encode input data.

module type RENDER = sig ... end

module Render (M : MONAD) (D : DECODABLE) : RENDER with type data = D.t and type output = string M.t

A functor that builds a render implementation for a given monad interface and a given decodable input type.

module RenderString (D : DECODABLE) : RENDER with type data = D.t and type output = string

A simpler version of Render that only requires a decodable module and outputs a string.

Printing templates as JavaScript modules.

type js_import

Information to import an external JavaScript function.

val js_import : module_path:string -> function_path:string -> js_import

val cjs : Stdlib.Format.formatter -> js_import compiled -> unit

Print a template as a CommonJS module.

val esm : Stdlib.Format.formatter -> js_import compiled -> unit

Print a template as an ECMAScript module.

Printing debug information.

val pp_typescheme : Stdlib.Format.formatter -> typescheme -> unit

Pretty-print a type scheme in Acutis syntax.

val pp_ast : Stdlib.Format.formatter -> parsed -> unit

Pretty-print a template's abstract syntax tree in S-expression syntax.

val pp_compiled : Stdlib.Format.formatter -> 'a compiled -> unit

Pretty-print a fully-compiled template in S-expression syntax.

val pp_instructions : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a compiled -> unit

Pretty-print runtime instructions in S-expression syntax.

val pp_js_import : Stdlib.Format.formatter -> js_import -> unit

Pretty-print JavaScript import data in S-expression syntax.