Module Acutis

module Ast : sig ... end

Define the untyped abstract syntax tree to be generated by the Parser.

module Compile : sig ... end

Orchestrate the Lexer, Parser, Typechecker, and Matching to produce the final template.

module Dagmap : sig ... end

A utility to help Typechecker and Compile enforce that templates form a directed acyclic graph.

module Data : sig ... end

The internal representation of runtime data.

module DataJson : sig ... end

Converts JSON strings to and from Data.t.

module Error : sig ... end

All of the error messages.

module Lexer : sig ... end

The lexer module generated by ocamllex.

module Loc : sig ... end

Define the locations of expressions in a source file.

module Map : sig ... end

Predefined map modules.

module Matching : sig ... end

Compile patterns into decision trees.

module Nonempty : sig ... end

Lists with at least one item.

module Parser : sig ... end

module ParserMessages : sig ... end

Turn Menhir errors into human-readable messages.

module Pp : sig ... end

Helpers for Format functions.

module PrintJs : sig ... end

Format compiled templates as self-contained JavaScript modules.

module Render : sig ... end

The main runtime for executing templates.

module Set : sig ... end

Predefined set modules.

module Sexp : sig ... end

Serialize internal data into S-expressions for debugging.

module Sync : sig ... end

An unboxed "monad" for synchronous code.

module Typechecker : sig ... end

Type-checks the untyped Ast.t and constructs a typed tree.

module Typescheme : sig ... end

Defines the types (integer, record, etc.).