Module Acutis.Ast

Define the untyped abstract syntax tree to be generated by the Parser.

module Dict : sig ... end

This is a Map.String which enforces that keys cannot change after they're set.

module Record : sig ... end

module Interface : sig ... end

type trim =

| No_trim

| Trim

type escape =

| No_escape

| Escape

type echo_format =

| Fmt_string

| Fmt_int

| Fmt_float

| Fmt_bool

The echo syntax is essentially a subset of the pattern syntax, one which only allows strings, variables, and record field access.

type echo =

| Echo_var of Loc.t * string

| Echo_string of Loc.t * string

| Echo_field of echo * string

type pat =

| Var of Loc.t * string

| Bool of Loc.t * int

| Int of Loc.t * int

| Float of Loc.t * float

| String of Loc.t * string

| Nullable of Loc.t * pat option

| Enum_string of Loc.t * string

| Enum_int of Loc.t * int

| List of Loc.t * pat list * pat option

| Tuple of Loc.t * pat list

| Record of Loc.t * pat Record.t

| Dict of Loc.t * pat Dict.t

| Block of Loc.t * t

| Field of Loc.t * pat * string

and node =

| Text of string * trim * trim

| Echo of (echo_format * echo) list * echo_format * echo * escape

| Match of Loc.t * pat Nonempty.t * case Nonempty.t

| Map_list of Loc.t * pat * case Nonempty.t

| Map_dict of Loc.t * pat * case Nonempty.t

| Component of Loc.t * string * string * pat Dict.t

| Interface of Loc.t * Interface.t

and case = {

pats : (Loc.t * pat Nonempty.t) Nonempty.t;

nodes : t;

}

and t = node list

val echo_format_to_sexp : echo_format -> Sexp.t

val escape_to_sexp : escape -> Sexp.t

val to_sexp : t -> Sexp.t