Module Acutis.Render

A functor that builds a render implementation for a given monad interface and a given decodable input type.

We need to use a functor because parameterized types, such as in MONAD, are not supported in first-class modules.

Parameters

module Monad : MONAD

module Data : DECODABLE

Signature

val apply : (Data.t -> string Monad.t) compiled -> Data.t -> string Monad.t

Apply data to a template and return the rendered output.