Module Render.Make

A functor that builds an implementation for a given monadic output type and a given data input type.

Output signature for Make.

Parameters

module M : MONAD

module D : DATA

Signature

type t = string M.t

type data = D.t

type component = data -> t Map.String.t -> t

val make : component Compile.t -> data -> t

Apply data to a template and return the rendered output.