Module Typechecker.Type

Manages the type definitions.

type row = [

| `Closed

| `Open

]

type int_bool =

| Not_bool

| Bool

type 'a sum = {

mutable cases : 'a;

mutable row : row;

}

Common fields shared by the enum and union types.

type ty =

| Unknown of row Stdlib.ref
(*
The row is for unification with sum types during destructuring.
*)

| Int

| Float

| String

| Nullable of t

| List of t

| Tuple of t list

| Record of record

| Dict of t * set_string Stdlib.ref
(*
The set tracks which keys have been used so they can build pattern-matching decision trees.
*)

| Enum_int of set_int sum * int_bool

| Enum_string of set_string sum

| Union_int of string * sum_union_int * int_bool

| Union_string of string * sum_union_string

and sum_union_int = record map_int sum

and sum_union_string = record map_string sum

and record = t map_string Stdlib.ref

and t = ty Stdlib.ref

type scheme = t map_string

val unknown : unit -> t

val int : unit -> t

val float : unit -> t

val string : unit -> t

val nullable : t -> t

val list : t -> t

val tuple : t list -> t

val record : record -> t

val dict : t -> t

val sum : 'a -> row -> 'a sum

val enum_int : set_int sum -> t

val enum_string : set_string sum -> t

val enum_false_and_true : unit -> t

val enum_false_only : unit -> t

val enum_true_only : unit -> t

val union_int : string -> sum_union_int -> t

val union_string : string -> sum_union_string -> t

val union_false_and_true : string -> f:record -> t:record -> t

val union_false_only : string -> record -> t

val union_true_only : string -> record -> t

val pp : Stdlib.Format.formatter -> t -> unit